SI-5730 hide constructors of sealed abstract classes in scaladoc
Sealed abstract classes (like `List`) have a primary constructor, public
by default. It can never be called by external code but it shows up in
the scaladoc as a nice `new List()` construtor...

If a class is only abstract, the constructor is still useful because
people can subclass and call it. If it is only sealed (i.e. effectively final),
then it is the normal constructor of a final class. But sealed *and*
abstract makes documenting the constructor useless.

This should remove the misleading constructors of `List`, `Double`,
`Option` and others from the scaladoc.