SI-7602 Avoid crash in LUBs with erroneous code
If a class contains a double defintion of a method that overrides
an interface method, LUBs could run into a spot where filtering
overloaded alternatives to those that match the interface method
fails to resolve to a single overload, which crashes the compiler.

This commit uses `filter` rather than `suchThat` to avoid the crash.