Add tests for flat classpath
Added:
- Tests comparing entries created using the old and the new classpath implementations.
- Tests for AggregateFlatClassPath which check whether it obtains correct entries from classpaths
specified in a constructor and whether it preserves the ordering in the case of repeated entries.
- Partest test generating various dependencies (directories, zips and jars with sources and class files)
and testing whether the compilation and further running an application works correctly when there are
these many types of entries specified as -classpath and -sourcepath.

Sample JUnit test that shows that all pieces of JUnit infrastructure work correctly now uses JUnit assert
as it should do from the beginning.