Add benchmarks to compare recursive and flat classpath implementations
The goal of these changes is to add possibility to:
- compare an efficiency and a content of both cp implementations (ClassPathImplComparator)
- examine the memory consumption by creating a lot of globals using specified classpath
(ClassPathMemoryConsumptionTester) - it can be considered as e.g. some approximation
of ScalaPresentationCompilers in Scala IDE when working with many projects

ClassPathMemoryConsumptionTester is placed in main (I mean not test) sources so thanks to
that it has properly, out of the box configured boot classpath etc. and it's easy to use it, e.g.:
scala scala.tools.nsc.ClassPathMemoryConsumptionTester -YclasspathImpl:<implementation_to_test> -cp <some_cp> -sourcepath <some_sp> -requiredInstances 50 SomeFileToCompile.scala
At the end it waits for the "exit" command so there can be used a profiler like JProfiler to
look how given implementation behaves.

Also set flat classpath implementation as a default one (e.g. to test on Jenkins)